+++ zoo/src/main/scala/com/intel/analytics/zoo/serving/PostProcessing.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.zoo.serving.utils
+package com.intel.analytics.zoo.serving
 
 import com.intel.analytics.bigdl.tensor.{Storage, Tensor}
+import com.intel.analytics.zoo.serving.utils.TensorUtils
 
 class PostProcessing(t: Tensor[Float]) {
   def tensorToNdArrayString(): String = {
