+++ zoo/src/test/scala/com/intel/analytics/zoo/serving/UtilsSpec.scala
@@ -16,21 +16,10 @@
 
 package com.intel.analytics.zoo.serving
 
-import com.intel.analytics.bigdl.tensor.{Storage, Tensor}
 import org.scalatest.{FlatSpec, Matchers}
 
-class PostProcessingSpec extends FlatSpec with Matchers {
-  "tensor to ndarray string" should "work" in {
-    val t1 = Tensor(Array(4.0f, 5, 3, 4), shape = Array(2, 2))
-    val ansStr = new PostProcessing(t1).tensorToNdArrayString()
-    val truthStr = "[[4.0,5.0],[3.0,4.0]]"
-    assert(ansStr == truthStr)
-  }
+class UtilsSpec extends FlatSpec with Matchers {
+  "s3, hdfs, remote model read" should "work" in {
 
-  "TopN filter" should "work properly" in {
-    val t1 = Tensor(Storage(Array(3.0f, 2, 1, 4, 5)))
-    val ansTensor = new PostProcessing(t1).topN(2)
-    val truthTensor = Tensor(Array(4.0f, 5, 3, 4), shape = Array(2, 2))
-    assert(ansTensor == truthTensor)
   }
 }
